restructure direcotories
