+++ src/components/Elements/BaseButton/index.ts
@@ -1,0 +1,4 @@
+import BaseButton from "./BaseButton.vue";
+export * from "./BaseButton.vue";
+
+export { BaseButton };
