+++ src/components/Elements/BaseDialog/index.ts
@@ -1,0 +1,4 @@
+import BaseDialog from "./BaseDialog.vue";
+export * from "./BaseDialog.vue";
+
+export { BaseDialog };
