+++ src/components/Elements/BaseDrawer/index.ts
@@ -1,0 +1,3 @@
+import BaseDrawer from "./BaseDrawer.vue";
+
+export { BaseDrawer };
