+++ src/components/Elements/BaseLink/index.ts
@@ -1,0 +1,3 @@
+import BaseLink from "./BaseLink.vue";
+
+export { BaseLink };
