+++ src/components/Elements/BaseSpinner/index.ts
@@ -1,0 +1,4 @@
+import BaseSpinner from "./BaseSpinner.vue";
+export * from "./BaseSpinner.vue";
+
+export { BaseSpinner };
