+++ src/components/Elements/BaseTable/index.ts
@@ -1,0 +1,3 @@
+import BaseTable from "./BaseTable.vue";
+
+export { BaseTable };
