+++ src/components/Elements/ConfirmationDialog/index.ts
@@ -1,0 +1,3 @@
+import ConfirmationDialog from "./ConfirmationDialog.vue";
+
+export { ConfirmationDialog };
