+++ src/components/Elements/MDPreview/index.ts
@@ -1,0 +1,3 @@
+import MDPreview from "./MDPreview.vue";
+
+export { MDPreview };
