+++ src/components/Elements/index.ts
@@ -1,19 +1,8 @@
-import BaseButton from "./BaseButton.vue";
-import BaseDialog from "./BaseDialog.vue";
-import BaseDrawer from "./BaseDrawer.vue";
-import BaseLink from "./BaseLink.vue";
-import BaseSpinner from "./BaseSpinner.vue";
-import BaseTable from "./BaseTable.vue";
-import ConfirmationDialog from "./ConfirmationDialog.vue";
-import MDPreview from "./MDPreview.vue";
-
-export {
-  BaseButton,
-  BaseDialog,
-  BaseDrawer,
-  BaseLink,
-  BaseSpinner,
-  BaseTable,
-  ConfirmationDialog,
-  MDPreview,
-};
+export * from "./BaseButton";
+export * from "./BaseDialog";
+export * from "./BaseDrawer";
+export * from "./BaseLink";
+export * from "./BaseSpinner";
+export * from "./BaseTable";
+export * from "./ConfirmationDialog";
+export * from "./MDPreview";
